Rename tensorflow module
